Created Exercises Component
Added a ExerciseCreate function to create an exercise.
Added a ExerciseEdit function to edit a pre-existing exercise.
Created an ExerciseForm to show the user forms to create an exercise.
Used CardDeck to display users exercises in a log.
Styled buttons and nav bar.
Created a Carousel to display images on the home page.